Add flags to input_verifier
These flags are currently not known by the verifier. As a result, this
causes a panic when such events are trying to get converted to the rust
object.

Add the missing flags in this CL.

Bug: 271455682
Test: manually enable event verification
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST

+++ libs/input/InputVerifier.cpp
@@ -44,7 +44,7 @@ Result<void> InputVerifier::processMovement(DeviceId deviceId, int32_t action,
     rust::Slice<const RustPointerProperties> properties{rpp.data(), rpp.size()};
     rust::String errorMessage =
             android::input::verifier::process_movement(*mVerifier, deviceId, action, properties,
-                                                       flags);
+                                                       static_cast<uint32_t>(flags));
     if (errorMessage.empty()) {
         return {};
     } else {
